requirements.txt
Flask==2.2.2
simplejson==3.17.6
slackclient==2.9.4
slackeventsapi==3.0.1


